vrank: add pfReport getter and split getters into getter.go
Add GetPfReport(blockNum) to VRankModule and implementation.

GetPfReport reads header round from Extra, returns empty report on round 0, and otherwise returns proposers for rounds [0, round).

Also refactor GetCfReport/GetPfReport out of handler.go into getter.go.
