+++ kaiax/vrank/impl/getter.go
@@ -1,0 +1,104 @@
+// Copyright 2026 The Kaia Authors
+// This file is part of the Kaia library.
+//
+// The Kaia library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The Kaia library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the Kaia library. If not, see <http://www.gnu.org/licenses/>.
+
+package impl
+
+import (
+	"bytes"
+	"slices"
+
+	"github.com/kaiachain/kaia/blockchain/types"
+	"github.com/kaiachain/kaia/common"
+	"github.com/kaiachain/kaia/kaiax/vrank"
+)
+
+// for building N-th header's VRank field, caller should query N-1 with the previous block's round.
+func (v *VRankModule) GetCfReport(blockNum, round uint64) (vrank.CfReport, error) {
+	// epoch header's VRank should be nil
+	if (blockNum+1)%vrankEpoch == 0 {
+		return vrank.CfReport{}, nil
+	}
+	if round > maxRound {
+		return nil, vrank.ErrRoundOutOfRange
+	}
+
+	// if I was not a validator for blockNum, I couldn't have collected cfReport for blockNum.
+	if !v.isValidator(blockNum) {
+		return vrank.CfReport{}, nil
+	}
+
+	vk := vrank.ViewKey{N: blockNum, R: uint8(round)}
+	prepreparedAt, expectedBlockHash, viewMap := v.collector.GetViewData(vk)
+	if prepreparedAt.IsZero() {
+		return nil, vrank.ErrPrepreparedTimeNotSet
+	}
+	candidates, err := v.Valset.GetCandidates(blockNum)
+	if err != nil || candidates == nil {
+		logger.Error("GetCandidates failed", "blockNum", blockNum)
+		return nil, vrank.ErrGetCandidateFailed
+	}
+	// safeCands = candidates who responded with expectedBlockHash before deadline.
+	safeCands := make(map[common.Address]struct{})
+	for sender, msgWithTime := range viewMap {
+		if !slices.Contains(candidates, sender) {
+			continue
+		}
+		if msgWithTime.Msg == nil || msgWithTime.Msg.BlockHash != expectedBlockHash {
+			continue
+		}
+		elapsed := msgWithTime.ReceivedAt.Sub(prepreparedAt).Milliseconds()
+		// early birds who sent before validator preprepared are safe
+		if elapsed <= candidatePrepareDeadlineMs {
+			safeCands[sender] = struct{}{}
+		}
+	}
+
+	// cfReport = candidates - safeCands = candidates who did not respond, responded after deadline, or lied (wrong BlockHash).
+	var cfReport vrank.CfReport
+	for _, addr := range candidates {
+		if _, ok := safeCands[addr]; !ok {
+			cfReport = append(cfReport, addr)
+		}
+	}
+	slices.SortFunc(cfReport, func(a, b common.Address) int { return bytes.Compare(a.Bytes(), b.Bytes()) })
+	return cfReport, nil
+}
+
+func (v *VRankModule) GetPfReport(blockNum uint64) ([]common.Address, error) {
+	header := v.Chain.GetHeaderByNumber(blockNum)
+	if header == nil {
+		return nil, vrank.ErrHeaderNotFound
+	}
+	if len(header.Extra) < types.IstanbulExtraVanity {
+		return nil, vrank.ErrHeaderExtraTooShort
+	}
+
+	round := uint64(header.Round())
+	if round == 0 {
+		return []common.Address{}, nil
+	}
+
+	pfReport := make([]common.Address, 0, round)
+	for r := uint64(0); r < round; r++ {
+		proposer, err := v.Valset.GetProposer(blockNum, r)
+		if err != nil {
+			return nil, err
+		}
+		pfReport = append(pfReport, proposer)
+	}
+
+	return pfReport, nil
+}
