+++ kaiax/vrank/interface.go
@@ -18,6 +18,7 @@ package vrank
 
 import (
 	"github.com/kaiachain/kaia/blockchain/types"
+	"github.com/kaiachain/kaia/common"
 	"github.com/kaiachain/kaia/consensus/istanbul"
 	"github.com/kaiachain/kaia/event"
 	"github.com/kaiachain/kaia/kaiax"
@@ -30,6 +31,7 @@ type VRankModule interface {
 	HandleVRankPreprepare(msg *VRankPreprepare) error
 	HandleVRankCandidate(msg *VRankCandidate) error
 	GetCfReport(blockNum, round uint64) (CfReport, error)
+	GetPfReport(blockNum uint64) ([]common.Address, error)
 
 	SubscribeVRank(sink chan<- *VRankBroadcastEvent) event.Subscription
 }
